Add common constant classes Remove unnecessary class methods
